Factor out an AppState struct
